requirements.txt
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
dnspython==2.6.1
Flask==3.0.3
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
pycparser==2.22
pymongo==4.9.1
python-dotenv==1.0.1
Werkzeug==3.0.4
